+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Minor: Add Demonic Brutus as a drop source to Beef for luck calculation. (#897)
 - Minor: Add `%PET%` template variable to inject the obtained pet name if known. (#899)
 - Minor: Add setting to adjust the base network timeout seconds. (#904)
+- Minor: Allow routing Dink notifications through a HTTPS proxy server. (#859)
 - Bugfix: Updated PoH regions after their locations were moved in-game. (#903)
 
 ## 1.12.1